pyproject.toml
[project]
name = "jako"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Taeho Kim", email = "dittos@gmail.com" }
]
dependencies = [
    "requests>=2.32.3",
    "beautifulsoup4>=4.12.3",
    "tiktoken>=0.8.0",
    "openai>=1.53.0",
    "pydantic>=2.9.2",
    "google-generativeai>=0.8.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ipykernel>=6.29.5",
    "pytest>=8.3.3",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/jako"]
